examples/pipeline/torch-cifar10/components/serving/requirements.txt
flask
numpy
opencv-python
torchvision
